Another day of scoring.
